requirements.txt

pandas>=2.1
numpy>=1.26
pyyaml>=6.0.1
scikit-learn>=1.3
matplotlib>=3.8
plotly>=5.22
streamlit>=1.34
statsmodels>=0.14
xgboost>=2.0
pytorch-lightning>=2.2
torchmetrics>=1.4
astral>=3.2
